Update README.md
Abstract: Efficient cell lysis is critical for extraction of DNA from difficult-to-lyse microorganisms such as Gram-positive bacteria and filamentous fungi. A bead-beating step is usually included in DNA extraction protocols to improve cell lysis. However, there is no consensus on the duration of bead-beating that is necessary for complete lysis of the microbial communities present in complex microbial ecosystems, but which will still maintain the integrity of DNA released from easy-to-lyse microbes. Another consideration is that most protocols are tailored to one particular target group of microbes, typically either bacteria or fungi, in a given sample matrix. In this study, we investigated the impact of five bead-beating durations (0, 3, 10, 15 and 20 minutes) during DNA extraction with the QIAamp® Fast DNA Stool Mini Kit, on the bacterial and fungal communities of single pig faecal and liquid feed samples, extracted in triplicate, with the objective of determining a suitable ‘catch-all’ method. Both sample types were subjected to the bead-beating durations in triplicate, followed by 16S (bacterial) and ITS2 (fungal) rDNA amplicon sequencing. The performance of the different bead-beating durations was assessed based on the quantity of total DNA extracted, alpha and beta-diversity analyses of the resultant microbial communities and differential abundance of bacterial and fungal taxa. Our results suggest that 20 minutes of bead-beating was most appropriate for maximising the lysis of difficult-to-lyse bacteria and fungi in both pig faeces and liquid feed, while minimising the negative impact on easier-to-lyse microbes. Total DNA yield increased with bead-beating duration for both sample types; however, the yield from faeces decreased after 20 minutes of bead-beating. Despite this, DESeq2 analysis indicated that changes in the differential abundances of the dominant taxa at this point were limited, which was supported by the Shannon diversity results. Maximising the bead-beating duration appeared to be necessary in order to obtain a representative profile of the Gram-positive bacteria, particularly in liquid feed, and of the filamentous fungi present in both sample types. However, considering the small sample size, along with the reliance on differential as opposed to absolute abundances to validate increases or decreases in taxa, a larger-scale study is necessary to verify the findings of the present study.